Moved TestModule into own file, added more test framework traits
